INT Extract common utils for introduce variable and introduce parameter.
